backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
paddlepaddle==3.2.2
paddleocr>=3.3.0
Pillow>=11.0.0
numpy>=2.1.0
anthropic==0.40.0
python-dotenv==1.0.1
